+++ packages/backend/src/api/resources/JSON.ts
@@ -474,25 +474,15 @@ export interface OrganizationInvitationAcceptedJSON extends OrganizationInvitati
   user_id: string;
 }
 
-/**
- * @interface
- */
+/** @inline */
 export interface PublicOrganizationDataJSON extends ClerkResourceJSON {
-  /**
-   * The name of the Organization.
-   */
+  /** The name of the Organization. */
   name: string;
-  /**
-   * The slug of the Organization.
-   */
+  /** The slug of the Organization. */
   slug: string;
-  /**
-   * Holds the default Organization profile image. Compatible with Clerk's [Image Optimization](https://clerk.com/docs/guides/development/image-optimization).
-   */
+  /** Holds the default Organization profile image. Compatible with Clerk's [Image Optimization](https://clerk.com/docs/guides/development/image-optimization). */
   image_url?: string;
-  /**
-   * Whether the Organization has a profile image.
-   */
+  /** Whether the Organization has a profile image. */
   has_image: boolean;
 }
 
@@ -768,6 +758,8 @@ export interface EnterpriseConnectionSamlConnectionJSON {
   idp_entity_id: string;
   idp_sso_url: string;
   idp_certificate: string;
+  idp_certificate_issued_at: number;
+  idp_certificate_expires_at: number;
   idp_metadata_url: string;
   idp_metadata: string;
   acs_url: string;
@@ -811,6 +803,8 @@ export interface SamlConnectionJSON extends ClerkResourceJSON {
   idp_entity_id: string;
   idp_sso_url: string;
   idp_certificate: string;
+  idp_certificate_issued_at: number;
+  idp_certificate_expires_at: number;
   idp_metadata_url: string;
   idp_metadata: string;
   acs_url: string;